add submenus
Add submenus
- EmployeeManagement_UI.py for all employee management functions
- PropertyManagement_UI.py for all property management functions
