+++ UI_Layer/SupervisorUI.py
@@ -3,10 +3,15 @@ import os
 import shutil
 import sys
 
-from UI_Layer.Validation import validate_email, validate_full_name, validate_ssn
+from UI_Layer.EmlpoyeeManagement_UI import EmployeeManagementUI
+from UI_Layer.PropertyManagement_UI import PropertyUI
+
+
 class SupervisorUI:
     def __init__(self, logic_wrapper: LogicWrapper):
         self.logic_wrapper = logic_wrapper
+        self.employee_management_ui = EmployeeManagementUI(logic_wrapper)
+        self.property_management_ui = PropertyUI(logic_wrapper)
 
     def clear_terminal(self):
         os.system('cls' if os.name == 'nt' else 'clear')
@@ -25,13 +30,8 @@ class SupervisorUI:
             print(c + "Air NaN Supervisor Portal".center(columns - 2, h) + c)
             print(d + "Welcome to the Supervisor Menu".center(columns - 2) + d)
             print(c + h * (columns - 2) + c)
-            print(d + " 1. List All Employees ".ljust(columns - 2) + d)
-            print(d + " 2. Register New Employee ".ljust(columns - 2) + d)
-            print(d + " 3. Update Employee Information ".ljust(columns - 2) + d)
-            print(d + " 4. List All Properties ".ljust(columns - 2) + d)
-            print(d + " 5. Add New Property ".ljust(columns - 2) + d)
-            print(d + " 6. Update Property Information ".ljust(columns - 2) + d)
-            print(d + " 7. Approve Maintenance Reports ".ljust(columns - 2) + d)
+            print(d + " 1. Employee Management ".ljust(columns - 2) + d)
+            print(d + " 2. Property Management ".ljust(columns - 2) + d)
             print(d + " b. Back to Login Menu ".ljust(columns - 2) + d)
             print(d + " q. Quit ".ljust(columns - 2) + d)
             print(c + h * (columns - 2) + c)
@@ -39,19 +39,9 @@ class SupervisorUI:
             choice = input("\nChoose an option: ").strip().lower()
 
             if choice == "1":
-                self.list_all_employees()
+                self.employee_management_ui.display_menu()
             elif choice == "2":
-                self.register_employee()
-            elif choice == "3":
-                self.update_employee_info()
-            elif choice == "4":
-                self.list_all_properties()
-            elif choice == "5":
-                self.add_new_property()
-            elif choice == "6":
-                self.update_property_info()
-            elif choice == "7":
-                self.approve_maintenance_reports()
+                self.property_management_ui.display_menu()
             elif choice == "b":
                 return
             elif choice == "q":
@@ -60,261 +50,3 @@ class SupervisorUI:
             else:
                 print("Invalid choice. Please try again.")
                 input("\nPress Enter to return to the menu.")
-
-    def list_all_employees(self):
-        self.clear_terminal()
-        columns, _ = self.get_terminal_size()
-        print("+".ljust(columns - 1, '-') + "+")
-        print("|" + " List All Employees ".center(columns - 2) + "|")
-        print("+".ljust(columns - 1, '-') + "+")
-        employees = self.logic_wrapper.list_employees()
-        if not employees:
-            print("No employees found.")
-        else:
-            headers = ["SSN", "Full Name", "Address", "Phone", "GSM", "Email", "Location", "Is Manager"]
-            col_widths = [max(len(str(getattr(emp, attr))) for emp in employees) for attr in ["ssn", "full_name", "address", "phone", "gsm", "email", "location", "is_manager"]]
-            col_widths = [max(len(header), width) for header, width in zip(headers, col_widths)]
-            row_format = "  |  ".join([f"{{:<{width}}}" for width in col_widths])
-            print(row_format.format(*headers))
-            print("-" * (columns - 2))
-            for emp in employees:
-                print(row_format.format(emp.ssn, emp.full_name, emp.address, emp.phone, emp.gsm, emp.email, emp.location, "True" if emp.is_manager else "False"))
-        input("\nPress Enter to return to the menu.")
-
-    def list_all_properties(self):
-        self.clear_terminal()
-        columns, _ = self.get_terminal_size()
-        print("+".ljust(columns - 1, '-') + "+")
-        print("|" + " List All Properties ".center(columns - 2) + "|")
-        print("+".ljust(columns - 1, '-') + "+")
-        properties = self.logic_wrapper.list_properties()
-        if not properties:
-            print("No properties found.")
-        else:
-            headers = ["Property ID", "Address", "Location", "Condition", "Manager", "Requires Maintenance"]
-            col_widths = [max(len(str(getattr(prop, attr))) for prop in properties) for attr in ["property_id", "address", "location", "property_condition", "manager", "requires_maintenance"]]
-            col_widths = [max(len(header), width) for header, width in zip(headers, col_widths)]
-            row_format = "  |  ".join([f"{{:<{width}}}" for width in col_widths])
-            print(row_format.format(*headers))
-            print("-" * (columns - 2))
-            for prop in properties:
-                print(row_format.format(prop.property_id, prop.address, prop.location, prop.property_condition, prop.manager, ", ".join(prop.requires_maintenance)))
-        input("\nPress Enter to return to the menu.")
-
-    def register_employee(self):
-        while True:
-            self.clear_terminal()
-            columns, _ = self.get_terminal_size()
-            print("+".ljust(columns - 1, '-') + "+")
-            print("|" + " Register New Employee ".center(columns - 2) + "|")
-            print("+".ljust(columns - 1, '-') + "+")
-            print("Enter 'b' at any prompt to cancel and go back to the previous menu.\n")
-            while True:
-                employee_details = {
-                    "ssn": input("Enter SSN: ").strip()
-                }
-                if employee_details["ssn"].lower() == 'b':
-                    return
-                if not validate_ssn(employee_details["ssn"]):
-                    print("Invalid SSN. It should be exactly 10 digits.")
-                    input("\nPress Enter to try again.")
-                else:
-                    break
-            employee_details.update({
-                "full_name": input("Enter Full Name: ").strip(),
-                "address": input("Enter Address: ").strip(),
-                "phone": input("Enter Phone: ").strip(),
-                "gsm": input("Enter GSM: ").strip(),
-                "location": input("Enter Location: ").strip(),
-                "is_manager": input("Is Manager (True/False): ").strip().lower() == 'true'
-            })
-            if any(value.lower() == 'b' for key, value in employee_details.items() if key != "is_manager"):
-                return
-            if not validate_full_name(employee_details["full_name"]):
-                print("Invalid Full Name. It should not be longer than 100 characters.")
-                input("\nPress Enter to try again.")
-                continue
-            while True:
-                employee_details["email"] = input("Enter Email: ").strip()
-                if employee_details["email"].lower() == 'b':
-                    return
-                if not validate_email(employee_details["email"]):
-                    print("Invalid Email. It should contain '@' and '.' and not be longer than 100 characters.")
-                    input("\nPress Enter to try again.")
-                else:
-                    break
-            result = self.logic_wrapper.register_employee(employee_details)
-            print(result)
-            print("\nEmployee registered successfully.")
-            self.list_all_employees()
-            break
-
-    def update_employee_info(self):
-        self.clear_terminal()
-        columns, _ = self.get_terminal_size()
-        print("+".ljust(columns - 1, '-') + "+")
-        print("|" + " Update Employee Information ".center(columns - 2) + "|")
-        print("+".ljust(columns - 1, '-') + "+")
-        print("Enter 'b' at any prompt to cancel and go back to the previous menu.\n")
-        employees = self.logic_wrapper.list_employees()
-        if not employees:
-            print("No employees found.")
-            input("\nPress Enter to return to the menu.")
-            return
-
-        headers = ["SSN", "Full Name", "Address", "Phone", "GSM", "Email", "Location", "Is Manager"]
-        col_widths = [max(len(str(getattr(emp, attr))) for emp in employees) for attr in ["ssn", "full_name", "address", "phone", "gsm", "email", "location", "is_manager"]]
-        col_widths = [max(len(header), width) for header, width in zip(headers, col_widths)]
-        row_format = "  |  ".join([f"{{:<{width}}}" for width in col_widths])
-        print(row_format.format(*headers))
-        print("-" * (columns - 2))
-        for emp in employees:
-            print(row_format.format(emp.ssn, emp.full_name, emp.address, emp.phone, emp.gsm, emp.email, emp.location, "True" if emp.is_manager else "False"))
-
-        ssn = input("\nEnter SSN of the employee to update: ").strip()
-        if ssn.lower() == 'b':
-            return
-        employee = next((emp for emp in employees if emp.ssn == ssn), None)
-        if not employee:
-            print(f"No employee found with SSN {ssn}")
-            input("\nPress Enter to return to the menu.")
-            return
-
-        print("\nCurrent Information:")
-        print(row_format.format(*headers))
-        print("-" * (columns - 2))
-        print(row_format.format(employee.ssn, employee.full_name, employee.address, employee.phone, employee.gsm, employee.email, employee.location, "True" if employee.is_manager else "False"))
-
-        field = input("\nEnter the field to update (full_name, address, phone, gsm, email, location, is_manager): ").strip()
-        if field.lower() == 'b':
-            return
-        if field == "email":
-            while True:
-                new_value = input(f"Enter new value for {field}: ").strip()
-                if new_value.lower() == 'b':
-                    return
-                if not validate_email(new_value):
-                    print("Invalid Email. It should contain '@' and '.' and not be longer than 100 characters.")
-                    input("\nPress Enter to try again.")
-                else:
-                    break
-        else:
-            new_value = input(f"Enter new value for {field}: ").strip()
-            if new_value.lower() == 'b':
-                return
-            if field == "ssn" and not validate_ssn(new_value):
-                print("Invalid SSN. It should be 10 digits.")
-                return
-            if field == "full_name" and not validate_full_name(new_value):
-                print("Invalid Full Name. It should not be longer than 100 characters.")
-                return
-        result = self.logic_wrapper.change_employee_info(ssn, field, new_value)
-        print(result)
-
-        # Fetch the updated employee information
-        updated_employee = self.logic_wrapper.search_employee_by_ssn(ssn)
-        print("\nUpdated Information:")
-        print(row_format.format(*headers))
-        print("-" * (columns - 2))
-        print(row_format.format(updated_employee.ssn, updated_employee.full_name, updated_employee.address, updated_employee.phone, updated_employee.gsm, updated_employee.email, updated_employee.location, "True" if updated_employee.is_manager else "False"))
-
-        input("\nPress Enter to return to the menu.")
-
-    def add_new_property(self):
-        self.clear_terminal()
-        columns, _ = self.get_terminal_size()
-        print("+".ljust(columns - 1, '-') + "+")
-        print("|" + " Add New Property ".center(columns - 2) + "|")
-        print("+".ljust(columns - 1, '-') + "+")
-        print("Enter 'b' at any prompt to cancel and go back to the previous menu.\n")
-        property_id = self.automatic_property_id()
-        property_details = {
-            "property_id": property_id
-        }
-        property_details.update({
-            "address": input("Enter Address: ").strip(),
-            "location": input("Enter Location: ").strip(),
-            "property_condition": input("Enter Property Condition: ").strip(),
-            "manager": input("Enter Manager: ").strip(),
-            "requires_maintenance": input("Enter Requires Maintenance (comma-separated): ").strip()
-        })
-        if any(value == 'b' for value in property_details.values()):
-            return
-        result = self.logic_wrapper.add_property(property_details)
-        print(result)
-        print("\nProperty added successfully.")
-        self.list_all_properties()
-
-    def update_property_info(self):
-        self.clear_terminal()
-        columns, _ = self.get_terminal_size()
-        print("+".ljust(columns - 1, '-') + "+")
-        print("|" + " Update Property Information ".center(columns - 2) + "|")
-        print("+".ljust(columns - 1, '-') + "+")
-        print("Enter 'b' at any prompt to cancel and go back to the previous menu.\n")
-        properties = self.logic_wrapper.list_properties()
-        if not properties:
-            print("No properties found.")
-            input("\nPress Enter to return to the menu.")
-            return
-
-        headers = ["Property ID", "Address", "Location", "Condition", "Manager", "Requires Maintenance"]
-        col_widths = [max(len(str(getattr(prop, attr))) for prop in properties) for attr in ["property_id", "address", "location", "property_condition", "manager", "requires_maintenance"]]
-        col_widths = [max(len(header), width) for header, width in zip(headers, col_widths)]
-        row_format = "  |  ".join([f"{{:<{width}}}" for width in col_widths])
-        print(row_format.format(*headers))
-        print("-" * (columns - 2))
-        for prop in properties:
-            print(row_format.format(prop.property_id, prop.address, prop.location, prop.property_condition, prop.manager, ", ".join(prop.requires_maintenance)))
-
-        property_id = input("\nEnter Property ID to update: ").strip()
-        if property_id.lower() == 'b':
-            return
-        property = next((prop for prop in properties if prop.property_id == property_id), None)
-        if not property:
-            print(f"No property found with ID {property_id}")
-            input("\nPress Enter to return to the menu.")
-            return
-
-        print("\nCurrent Information:")
-        print(row_format.format(*headers))
-        print("-" * (columns - 2))
-        print(row_format.format(property.property_id, property.address, property.location, property.property_condition, property.manager, ", ".join(property.requires_maintenance)))
-
-        field = input("\nEnter the field to update (address, location, property_condition, manager, requires_maintenance): ").strip()
-        if field.lower() == 'b':
-            return
-        new_value = input(f"Enter new value for {field}: ").strip()
-        if new_value.lower() == 'b':
-            return
-        updated_details = {field: new_value}
-        result = self.logic_wrapper.update_property(property_id, updated_details)
-        print(result)
-
-        # Fetch the updated property information
-        updated_property = self.logic_wrapper.get_property_by_id(property_id)
-        print("\nUpdated Information:")
-        print(row_format.format(*headers))
-        print("-" * (columns - 2))
-        print(row_format.format(updated_property.property_id, updated_property.address, updated_property.location, updated_property.property_condition, updated_property.manager, ", ".join(updated_property.requires_maintenance)))
-
-        input("\nPress Enter to return to the menu.")
-
-    def approve_maintenance_reports(self):
-        self.clear_terminal()
-        columns, _ = self.get_terminal_size()
-        print("+".ljust(columns - 1, '-') + "+")
-        print("|" + " Approve Maintenance Reports ".center(columns - 2) + "|")
-        print("+".ljust(columns - 1, '-') + "+")
-        # Implementation for approving maintenance reports
-        input("\nPress Enter to return to the menu.")
-
-    #Possibly should be in another layer?.
-    def automatic_property_id(self):
-        """
-        Gets the latest property ID and give it plus 1.
-        """
-        properties = self.logic_wrapper.list_properties()
-        latest_property = properties[-1]
-        latest_id = int(latest_property.property_id)
-        return latest_id + 1
